+++ src/account-payment-handler.ts
@@ -16,11 +16,11 @@ export class AccountPaymentHandler {
     await this.client.init();
   }
 
-  async sendCoins(recipient: Account | string, amount: number) {
-    await this.client.sendCoins(new TransferDetails(this.account, recipient, amount));
+  async sendCoins(recipient: Account | string, amount: number, memo = "") {
+    return await this.client.sendCoins(new TransferDetails(this.account, recipient, amount, memo));
   }
 
   async sendBulkTransactions(transactions: Transaction[]) {
-    await this.client.sendBulkTransactions(this.account, transactions);
+    return await this.client.sendBulkTransactions(this.account, transactions);
   }
 }