+++ tests/account.test.js
@@ -1,4 +1,4 @@
-const { Account } = require("../");
+const { Account } = require("../dist");
 
 describe("Account", () => {
   const defaultAccount = {
@@ -46,6 +46,15 @@ describe("Account", () => {
     expect(Account.isValidPair(defaultAccount.accountNumber, defaultAccount.signingKey)).toBeFalsy();
   });
 
+  it("isValidPair doesn't throw errors", () => {
+    const results = [
+      Account.isValidPair("asdf", "asdf"),
+      Account.isValidPair(defaultAccount.signingKey, "asdf"),
+      Account.isValidPair("asdf", defaultAccount.accountNumber),
+    ];
+    expect(results.every((val) => typeof val === "boolean")).toBeTruthy();
+  });
+
   it("createSignature(message)", () => {
     const account = createDefaultAccount();
     assertAccountBasics(account);
@@ -71,6 +80,32 @@ describe("Account", () => {
     });
   });
 
+  it("verifySignature(message: string, signature: string, accountNumber: string)", () => {
+    const account = createDefaultAccount();
+    const message = JSON.stringify({
+      balance_key: null,
+      txs: [
+        {
+          amount: 1,
+          fee: "PRIMARY_VALIDATOR",
+          recipient: "4afb3eaad999e4c073be0fbde86b76f9370d53b398b9cab9d760825709a1d6b3",
+        },
+        {
+          amount: 100,
+          memo: "Deposit",
+          recipient: "718b3c72e1e520479b9f9a2a582f280e549732c97c44de22c5b2a4e443154342",
+        },
+        { amount: 1, fee: "BANK", recipient: "e3a94381f8db207ddad931391886d611d6f4c060d0db2b0e373738e2f4db96d6" },
+      ],
+    });
+
+    const signature = account.createSignature(message);
+
+    expect(Account.verifySignature(message, signature, account.accountNumberHex)).toBeTruthy();
+    expect(Account.verifySignature("Wrong Message", signature, account.accountNumberHex)).toBeFalsy();
+    expect(Account.verifySignature(message, "Wrong Signature", account.accountNumberHex)).toBeFalsy();
+    expect(Account.verifySignature(message, signature, "Wrong Account Number")).toBeFalsy();
+  });
   // TODO: createBlockData
 
   // TODO: createBlockMessage