+++ CHANGELOG.md
@@ -2,6 +2,50 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0-alpha.5](https://github.com/thenewboston-developers/thenewboston-js/compare/v1.1.0-alpha.4...v1.1.0-alpha.5) (2021-06-30)
+
+
+### Features
+
+* Verify Signatures ([#164](https://github.com/thenewboston-developers/thenewboston-js/issues/164)) ([267e89b](https://github.com/thenewboston-developers/thenewboston-js/commit/267e89bae70466c8ff2550fdc0aa5e7005d30389))
+
+## [1.1.0-alpha.4](https://github.com/thenewboston-developers/thenewboston-js/compare/v1.1.0-alpha.3...v1.1.0-alpha.4) (2021-04-29)
+
+
+### Features
+
+* Replaced type with enum for types with multiple values ([#157](https://github.com/thenewboston-developers/thenewboston-js/issues/157)) ([cab172b](https://github.com/thenewboston-developers/thenewboston-js/commit/cab172be7a8d87713fa88ac601e916f99715c71d))
+
+
+### Bug Fixes
+
+* `Account.isValidPair` throwing on invalid seeds ([#152](https://github.com/thenewboston-developers/thenewboston-js/issues/152)) ([1a18fad](https://github.com/thenewboston-developers/thenewboston-js/commit/1a18fad5e143658e66b2beb01829fbd1c99780d1)), closes [#151](https://github.com/thenewboston-developers/thenewboston-js/issues/151)
+
+## [1.1.0-alpha.3](https://github.com/thenewboston-developers/thenewboston-js/compare/v1.1.0-alpha.2...v1.1.0-alpha.3) (2021-04-12)
+
+
+### Features
+
+* Added support for memo ([#140](https://github.com/thenewboston-developers/thenewboston-js/issues/140)) ([ad5ed54](https://github.com/thenewboston-developers/thenewboston-js/commit/ad5ed54efc1f09b0ec2525fc0255ab39ab1df61d))
+
+## [1.1.0-alpha.2](https://github.com/thenewboston-developers/thenewboston-js/compare/v1.1.0-alpha.1...v1.1.0-alpha.2) (2021-04-10)
+
+
+### Bug Fixes
+
+* Replaced startCrawl and stopCrawl return types ([#136](https://github.com/thenewboston-developers/thenewboston-js/issues/136)) ([9f05092](https://github.com/thenewboston-developers/thenewboston-js/commit/9f05092d97e172eef5e1736a266795cfdc721632))
+
+## [1.1.0-alpha.1](https://github.com/thenewboston-developers/thenewboston-js/compare/v1.1.0-alpha.0...v1.1.0-alpha.1) (2021-04-09)
+
+### Features
+
+* feat: Get Methods for Single nodes ([#121](https://github.com/thenewboston-developers/thenewboston-js/issues/121)) ([be17ed9](https://github.com/thenewboston-developers/thenewboston-js/commit/be17ed9d1f50d6a439bbfd4efb1506b324b384e0))
+* feat: added updateBankTrust, updateValidatorTrust ([#118](https://github.com/thenewboston-developers/thenewboston-js/issues/118)) ([32e1e1c](https://github.com/thenewboston-developers/thenewboston-js/commit/32e1e1c2afd549bfde80ca4fccc29da55ab5ee32))
+
+### Bug Fixes
+
+* Updated CV and Pv response types ([#133](https://github.com/thenewboston-developers/thenewboston-js/issues/133)) ([7191322](https://github.com/thenewboston-developers/thenewboston-js/commit/7191322e956bc4e2ac4bf47c45a4a075917e8c0d))
+
 ## [1.1.0-alpha.0](https://github.com/thenewboston-developers/thenewboston-js/compare/v1.0.3...v1.1.0-alpha.0) (2021-04-01)